Backend/requirements.txt
torch
torchvision
torchaudio
numpy
pandas
opencv-python
pillow
requests
matplotlib
tqdm
scipy
seaborn
uvicorn
fastapi
